testing

+++ Coursework/comments-func.php
@@ -56,10 +56,10 @@ function retrieveComments($link) { // to retrieve comments from the database, li
 }
 
 function deletePosts($link) {
-    if (isset($_POST['deletePost'])) { // unless button is pressed shouldn't run code below
+    if (isset($_POST['deletePost'])) { // unless delete button is pressed shouldn't run code below
         $cid = $_POST['cid'];
 
-        $sql = "delete from tbl_comments where cid='$cid'";
+        $sql = "delete from tbl_comments where cid='$cid' and uid='". $_SESSION['id'] ."'";
         $result = $link->query($sql); // variable to store connection to use query on sql variable about with update statement above
         header("location: comments.php");
     }
